New notion of subspawnpoints. They replace the default wingmen formation if given:
 <Spawn x="10" y="0" xdir="1" ydir="0">
   <Spawn x="0" y="10" xdir="0" ydir="1"/>
   <Spawn x="-10" y="0" xdir="-1" ydir="0"/>
   <Spawn x="0" y="-10" xdir="0" ydir="-1"/>
 </Spawn>

This would make the first four players of the team that gets this spawnpoint assigned to it point away from each other. If there are more than four players in that team they'll get spawned according to the standard wingmen formation (using the location of the root spawn point as base).

The direction of the subspawnpoints defaults to the direction of the root spawn point if not given.