Updated example-1.0.0.aamap.xml map file and added new DTD: map-0.2.9_styctap_v1.5.dtd It now contains the zone types: blast and koh; These had been forgotten in previous additions.
Modified the RACE_SAFE_ANGLES code in gCycle. The problem had been occuring due to the weird calculations the script was performing for straight angles like 90, 180 and 270. So, I have coded a precission check which will verify whether the direction of x and y are zero or not. Then it will re-append the value which was expected.