Make binreloc work again
Apparenly, a recent-ish change in Linux binary loading now has
constant strings in a segment that is marked differently in
/proc/self/map, a post-memory-relocation readonly area, I gather.
It is marked with r--p. Adapt binreloc to also accept such
segments.

Related to https://gitlab.com/armagetronad/armagetronad/issues/6,
should reduce the number of cases the workaround needs to kick in.