Fix order of deployments
First: scp (sourceforge or staging area) and launchpad
Second: Steam
Third: Zero Install and download page
Fourth: PPA

Rationale:
The first wave can be safely rerun and does not directly affect users.

The first and second wave deployments depend on external
services and may fail without fault on our part, and if that happens,
we can only wait and retry.

The third wave only interacts with services under our full control, so
if they fail, we can fix the problems and try again.

The fourth wave is fire and forget, we don't get to react to errors anyway,
but reruns cause mail spam.