Improve and clean up gZone::InteractWith a little bit.
Should improve load with heavy zone usage at least a tiny bit. We're no longer checking if dead players are close to a zone or not and we're no longer calculating the same distance figure over and over again.

Comment out the ONLINE_ZONE thing as well; I'm not quite sure what the point is with the other ladder log events surrounding spawned zones and it would never work in its current state. I'll decide later what to actually do with it though.