Do not send team change messages for other players
In ePlayerNetID::RemoveFromGame, only send team change messages if
the player still is on a team. They are, at best, redundant; at worst,
they are illegal, because the function is called on the client on
destruction of ePlayerNetIDs that used to belong to other clients
and that have been taken over by now.

Fixes https://gitlab.com/armagetronad/armagetronad/-/issues/98
the client part of it.