Potential client crashfix: Reduce the max packet size when destroying network objects. (down to 500 instead of 1000)
As per my understanding:
Large packets (a datagram over 576 bytes) should be split, and eventually reassembled. They are, apparently, also allowed to simply be dropped instead for any reason, which some routers do.

In this particular case, there is no real reason to rely on that behavior as we were already limiting the size.