Make abandoning direct SF downloads permanent
While it turned out sourceforge was probably not the reason for the
certificate failures causing failed deployments and flatpak builds
(rather, it was outdated ca information on the builders),
the site is slow and full of ads. An if-not-found-redirect to SF
ensures we can recklessly clean up our files anyway.